requirements.txt
Django==2.2.2
django-debug-toolbar==1.11
djangorestframework==3.9.4
djangorestframework-stubs==0.4.2
pytz==2019.1
sqlparse==0.3.0
